minor edit to cleanup script

+++ cleanup.py
@@ -23,6 +23,7 @@ for compute_target in ws.compute_targets:
             pass
     elif (
         "dask-ct" in compute_target
+        and ws.compute_targets[compute_target].get_status() in ["Succeeded"]
         and len(ws.compute_targets[compute_target].list_nodes()) == 0
     ):
         ws.compute_targets[compute_target].delete()
